requirements.txt
requests
beautifulsoup4
python-whois
dnspython
pygeoip
reportlab